refactor: separate all frontend components into apps
