+++ frontend/src/apps/inventory/hooks/index.tsx
@@ -1,1 +1,2 @@
 export { useProducts } from './useProducts';
+export { useInventory } from './useInventory';
