+++ frontend/src/apps/inventory/hooks/useInventory.ts
