+++ frontend/src/apps/inventory/pages/ArchivedProducts.tsx
@@ -3,7 +3,7 @@ import { Link } from "react-router-dom";
 import { ArrowLeft } from "lucide-react";
 import ProductTable from "../components/ProductTable";
 import { fetchArchivedProducts, toggleProductArchive } from "../services/productService";
-import { type Product } from "../types/product";
+import { type Product } from "../../../types/product";
 import toast from 'react-hot-toast';
