+++ frontend/src/apps/sales/index.tsx
@@ -1,4 +1,4 @@
-export { SalesDashboard, SalesList, SaleForm, SaleDetail } from './pages';
+export { InvoicesList, InvoiceDetail, SalesDashboard, SalesList, SaleForm, SaleDetail } from './pages';
 export { useSales, useSale, useAvailableProducts, useSalesStats, useSalesCalculations } from './hooks';
 export { SalesService, InvoiceService, InventoryService } from './services';
 export type {
@@ -7,4 +7,4 @@ export type {
   SaleItem,
   SalesFilters,
   SalesStats
-} from '@/types/product';
+} from './types_sales';
