+++ frontend/src/apps/sales/pages/index.tsx
@@ -1,3 +1,5 @@
+export { InvoiceDetail } from './InvoiceDetail';
+export { InvoicesList } from './InvoicesList';
 export { SalesDashboard } from './SalesDashboard';
 export { SalesList } from './SalesList';
 export { SaleForm } from './SaleForm';
