+++ frontend/src/apps/sales/types_sales.ts
@@ -1,72 +1,3 @@
-// TODO (optional) generate types dynamically from backend models
-
-export interface Product {
-  id: number;
-  type: number;
-  type_name_ar: string;
-  type_name_en: string;
-  brand: number | null;
-  brand_name_ar: string | null;
-  brand_name_en: string | null;
-  material: number | null;
-  material_name_ar: string | null;
-  material_name_en: string | null;
-  cost_price: number;
-  selling_price: number;
-  size: string;
-  weight: number | null;
-  tags: string;
-  tags_list: string[];
-  profit: number;
-  profit_percentage: number;
-  created_at: string;
-  updated_at: string;
-  created_by: number;
-  available_stock: number;
-  is_low_stock: boolean;
-  is_active: boolean;
-}
-
-export interface ProductType {
-  id: number;
-  name_en: string;
-  name_ar: string;
-  created_at: string;
-}
-
-export interface ProductBrand {
-  id: number;
-  name_en: string;
-  name_ar: string;
-  created_at: string;
-}
-
-export interface ProductMaterial {
-  id: number;
-  name_en: string;
-  name_ar: string;
-  created_at: string;
-}
-
-export type Filters = {
-  search: string;
-  type: string;
-  brand: string;
-  material: string;
-};
-
-export interface Inventory {
-  id: number;
-  product: number;
-  product_name: string;
-  product_type: string;
-  quantity_in_stock: number;
-  minimum_stock_level: number;
-  last_updated: string;
-  is_low_stock: boolean;
-  is_out_of_stock: boolean;
-}
-
 export interface SaleItem {
   id?: number;
   product: number;
@@ -124,32 +55,6 @@ export interface SaleListItem {
   profit_percentage: number;
 }
 
-export interface Invoice {
-  id?: number;
-  invoice_number?: string;
-  sale: number;
-  invoice_date?: string;
-  due_date?: string | null;
-  company_name: string;
-  company_address: string;
-  company_phone: string;
-  company_email: string;
-  is_printed: boolean;
-  printed_at?: string | null;
-  created_at?: string;
-  updated_at?: string;
-  sale_details?: Sale;
-}
-
-export interface InvoiceListItem {
-  id: number;
-  invoice_number: string;
-  invoice_date: string;
-  customer_name: string;
-  final_total: number;
-  is_printed: boolean;
-}
-
 export interface SalesStats {
   overview: {
     total_sales: number;
@@ -193,9 +98,28 @@ export interface SalesFilters {
   date_to: string;
 }
 
-export interface ApiResponse<T> {
-  results?: T[];
-  count?: number;
-  next?: string | null;
-  previous?: string | null;
+export interface Invoice {
+  id?: number;
+  invoice_number?: string;
+  sale: number;
+  invoice_date?: string;
+  due_date?: string | null;
+  company_name: string;
+  company_address: string;
+  company_phone: string;
+  company_email: string;
+  is_printed: boolean;
+  printed_at?: string | null;
+  created_at?: string;
+  updated_at?: string;
+  sale_details?: Sale;
+}
+
+export interface InvoiceListItem {
+  id: number;
+  invoice_number: string;
+  invoice_date: string;
+  customer_name: string;
+  final_total: number;
+  is_printed: boolean;
 }
